requirements.txt
asgiref~=3.3.1
attrs~=20.1.0
beautifulsoup4==4.9.3
Django~=3.0.8
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
iniconfig==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz~=2019.3
six~=1.14.0
soupsieve==2.0.1
sqlparse~=0.3.1
toml~=0.10.1
colorama~=0.4.3
numpy~=1.19.1
pip~=20.2.3
keyring~=18.0.1
PyYAML~=5.3.1
Jinja2~=2.11.2
setuptools~=49.6.0
django-model-utils~=4.0.0
django-filter~=2.4.0